cli/requirements.txt
click>=8.1.0
rich>=13.0.0
pydantic>=2.0.0
pyyaml>=6.0
httpx>=0.25.0
watchdog>=3.0.0
jinja2>=3.1.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# ML packages with compatible versions
torch>=2.0.0,<2.2.0
transformers>=4.41.0,<4.50.0
datasets>=2.14.0
tokenizers>=0.20.0,<0.21.0
accelerate>=0.20.0

# Optional - only install if training locally
# unsloth[cu121-torch240] @ git+https://github.com/unslothai/unsloth.git
# trl>=0.7.0
# peft>=0.5.0
